requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.3
django-celery-beat==2.6.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
kombu==5.3.6
pathlib==1.0.1
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg2==2.9.9
python-crontab==3.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
wheel==0.43.0
